Perf improvement for Seq.windowed, adapted from suggestion originally by Jon Harrop
For larger windows, Array.Copy is *much* faster than Array.init.  For smaller windows, Array.init is faster.  Threshold of 32 is based on testing with various data types on various hardware - goal is to err on the high side for the threshold, so that at least nobody encounters a perf regression due to this change.

Also added to the unit tests for Seq.windowed, which were a bit light. (changeset 1258476)