Fix for #69 - Invalid code generated when calling C# method with optional nullable args
Old codegen was improperly emitting default Nullable() for all nullable optional args that were not specified by the developer.  This was correct when the default value was in fact "null" (i.e. Nullable()), but was incorrect when default value contains a real value (i.e. Nullable(42)).  This change now detects when default value is a System.Nullable, and emits proper call to Nullable(xyz) constructor when required.  This change is based on suggestion form Vlad in email a while back.

Also noticed (from comment in the code) that VB allows byref optional args.  Upon investigation, these were also being handled incorrectly, so I refactored a bit in order to handle those. (changeset 1280456)