Codeplex issue: 132 - FSI Shadowcopy -- causes a significant degrade in the fsi first execute time
if you type
 1+1;;
 2+2;;
 there is a perhaps 1 second delay between hitting return and seeing the answer when shadow copy is enabled this delay does not occur when shadow copy is not enabled.
 subsequent statements to the first are not delayed significantly.

The Fix is to enable LoaderOptimisation.MultiDomainHost -- this causes the clr to use the ngened images for assemblies that are signed, loaded from the gac and have corresponding ngen images.   In the issue scenario, when enabling shadowcopy, the appdomain loader unable to load the ngen images in the appdomain doing the work.  This is a known limitation of AppDomains.  The solution is to make the assemblies shared accross appdomains

I did some testing and for a script similar to the 1+1;; in the issue report I ran fsi with and without shadowcopyreferences enabled, with and without MultiDomainHost enabled a few hundred times the table below is the average times of each set of runs.  In all cases Fsharp files were ngened and loaded from their standard VS install locations.

		Without MultiDomainHost			With MultiDomainHost
		/shadowcopy-	/shadowcopy+		/shadowcopy-	/shadowcopy+
		56840.58411	118600.1729		61529.93774	71692.79377

As you can see the impact of shadowcpy is much smaller, close to without.  The extra work here is due to creating a new appdomain so the devel it can be unloaded.

MultiDomainHost causes static fields to be indirected and so there is a small degrade on with/without.  The timings were done on a dualcore VM on a server that was not doing any other work.

 (changeset 1338154)