Add bug template to CONTRIBUTING.md
We've been getting bug reports which don't include good repro steps.  Unfortunately GitHub doesn't seem to give the ability to give a predefined template for issue reports, which is a real shame, since having a predefined template really helps improve the quality of reports.

As a partial step in that direction we can at least add something to CONTRIBUTING.md to explain what we expect minimally.

closes #75