Fixes #157 -- should not allow Framework 4 / F# 3.1 combination in project properties
Improves FSharp.Core selection given a specific .Net Framework.

When a new target framework is selected if the currently selected fsharp.core is not compatible
with the newly selected .net framework then it is selected according to the following:

.NetFrameworks 2.0, 3.0 and 3.5 select fsharp.core, v 2.3.0.0
.NetFramework 4.0 selects fsharp.core v 4.3.0.0
.NetFramework 4.5, 4.51, 4.52 and 4.6 selects fsharp.core 4.4.0.0

If a compatible fsharp.core is selected when a new .netframework is selected then the selection remains unchanged.
.NetFramework 4.5, 4.51, 4.52 and 4.6 are compatible with fsharp.core and 4.3.1.0, 4.4.0.0

Note:
1.  Internal build "copydependencies.proj" needs to be updated with a new SupportedRuntimes.xml location.
2.  OSS supportedruntimes.xml needs to be manually copied to the %ProgramFiles(x86)%\Microsoft SDKs\F#\4.0\Framework\v4.0 directory.
    until OSS vsix installer is updated to integrate the compiler into VS.