Fixes #9, Internal error in FSI: FS0192: binding null type in envBindTypeRef #9 Fixes #10, internal error: binding null type in envBindTypeRef
Both issues were caused by the same underlying issue.  RefEmit automagically applies escaping to names that contain a ','  So ``,`` becomes ``\,``.  When we tried to create a ref to a nested type we didn't add the ',' escaping so we couldnot bind to the created type.

The fix is very straightforward, change BasicQualifiedName to correctly escape names with comma's.