'for i in expr do body' optimization..
Previously 'for i in expr do body' only had optimization
when the type of expr was an 1D-array.

'for i in expr do body' now has optimizations for when expr
is of type 'string' and 'List`1'.

These optimizations increases the performance of the for
expression but also reduces the number of objects created.

latkin provided the following fixes:
1. Adapting tests to work with core\portable and core\netcore
2. Loop item over strings sometimes uses object, not char
3. Adjustments to debug ranges so that debug stepping behavior is consistent/unchanged
4. Updating codegen tests to represent corrected debug ranges