Fix fsharpqa CodeGen/EmittedIL tests for exception filter handlers.
Fixed the fsharpqa CodeGen/EmittedIL tests so they work when the
compiler generates exception filter handlers (which is the new default).
Affected tests have been copied (before modifying them) to create new
test cases which check that the old behavior is preserved when the
--exnfilters- compiler flag is specified.

Some of the CodeGen/EmittedIL tests included baselines for both .NET 2.0
and .NET 4.0, but only the .NET 4.0 tests were actually specified in the
env.lst files. In those cases new baselines were only created for
.NET 4.0.