Change pdbClose to test if files are locked before inducing GCs
closes #346

The original code triggers 6 full GCs, the time is a multiple of the size of the output binary since all of the IL/metadata structures are still on the heap at this point. This takes ~600ms on my test project (of a 10s total compile).

Changed to perform a check if the file is locked before inducing GC. I have been running the change and notice that occasionally one iteration of the loop is required, so the detection is working. If you remove all the code you will get a compiler error in the case that the file is locked, so this will not cause a silent regression.