The FSharp.CompilerService has encountered scenarios where SplitNameSpace caching requires thread safety.
in il.fs
    memoizeNamespaceTable
    memoizeNamespaceArrayTable

Are global mutable state, this change makes them both private, and thread safe.

It should be noted that they are both grow only caches, and thus will perform like a memory leak
when used in long running process that compile many different projects.  Although for the current use cases they are O.K

The implementation locks on write only.  All reads are free threaded.  If two or more threads collide on the same key,
the cache will contain the last value added, but all threads will compute the same content.
This is fine, since the cache contains an array created by splitting the key, into it's elements.