Explicitly set custom F# VS settings if they are missing
On package load, check if any of the custom F# settings are somehow missing
from the settings store.  Explicitly set them to their desired defaults if
they are indeed missing.

I've added this to the language service package since this is guaranteed
to load any time the project system package is loaded, but the converse is
not true (e.g. open loose F# script file without opening a solution).

fixes #199