Explicitly set custom F# VS settings if they are missing
On package load, check if any of the custom F# settings are somehow missing
from the settings store.  Explicitly set them to their desired defaults if
they are indeed missing.

Added to the language service package since this is guaranteed
to load any time the project system package is loaded, but the converse is
not true (e.g. open loose F# script file without opening a solution).

fixes #199
closes #491

commit da9f29aed5cf4911f7ec57f0632b7e9ed78fe6bc
Author: latkin <latkin@microsoft.com>
Date:   Wed Jun 10 09:33:28 2015 -0700

    Don't create unused binding in VS 12.0

commit 92c89198a0bdb937a70b0078576df17d854d2f65
Author: latkin <latkin@microsoft.com>
Date:   Tue Jun 9 18:55:54 2015 -0700

    APIs used are new for 14.0, omit them for 12.0 builds

commit a662e18d90fa968b9102c503b63f1dda9c4e9f72
Author: latkin <latkin@microsoft.com>
Date:   Tue Jun 9 17:28:22 2015 -0700

    Code review feedback, and add clickable URLs setting

commit b6697a7c24eceb4413987d0453be9cbea4ebb5d9
Author: latkin <latkin@microsoft.com>
Date:   Mon Jun 8 18:13:35 2015 -0700

    Explicitly set custom F# VS settings if they are missing