added range check to readBlobHeapUncached
Obfuscated assemblies can have a bogus extra reference to mscorlib which
has incorrect index in Blob heap for HashValue - index value was out
of range of the heap.I guess the assumption was that compiler will not
try to realize the whole set of information for this AssemblyRef entry
but decompiler will try to do. The problem was that F# compiler is
eagerly trying to populate information about referenced assemblies for the
given assembly so it crashes during the attempt to read this malformed entry.

fixes https://github.com/Microsoft/visualfsharp/issues/517
closes https://github.com/Microsoft/visualfsharp/pull/519

commit 326a7fc8fe346e1ee65fad9f60b3d7b84cf18a42
Author: Vladimir Matveev <vladima@microsoft.com>
Date:   Wed Jul 1 00:27:11 2015 -0700

    added tests

commit 1d8dd07e7652fd20cce3300bf7099919a1d900a2
Author: Vladimir Matveev <vladima@microsoft.com>
Date:   Mon Jun 29 00:14:27 2015 -0700

    fix incorrect condition

commit 0c0c69624da25ca3453b3fc84207a86e75cb015d
Author: Vladimir Matveev <vladima@microsoft.com>
Date:   Sun Jun 28 23:30:38 2015 -0700

    added range check to readBlobHeapUncached