Add a lighter-weight API for checking if an assembly contains a type provider
Old situation for checking if a project had referenced a type provider:
  - Language service had somewhat unreliable, hard-to-predict way of detecting TPs
    - Raise event when LS compiler hits a TP during background compile (hard to know when that will happen)
    - Rely on global mutable state in compiler to track current source file being processed -> use that to tie back to current project
  - Project system had a deterministic and accurate (but very slow) way of detecting TPs
    - Run through the beginning of the full compile pipeline, but stop after references are imported

New API is fairly lightweight, and targeted at a single file - no need to launch full compile machinery. It simply
opens an IL reader to a specified assembly and checks for TypeProviderAssemblyAttribute. This looks to be fast enough
for on-demand use in the IDE tools.