Second half of fix to https://github.com/Microsoft/visualfsharp/issues/106
FSharp.Compiler.Service contained a more complete fix for
https://github.com/Microsoft/visualfsharp/issues/106, which relates to the
memory leaks that can arise if type provider instance objects are kept live.
Historically the memory leaks have happened if a type provider registers
itself to listen to file system events (or other notifications), but for some
reason doesn't disconnect those listeners when the type provider is disposed.
This keeps the type provider instance alive - which is a bug in the type
provider - but even worse it can keep '00s MB of data associated with the
compiler state alive as well.

The missing part of the fix is to disconnect the "TypeProviderConfig" callback
when the type providers are disposed, so that even if a type provider instance
hangs on to a reference to the TypeProviderConfig, that doesn't cause the entire
object graph for the compilation session to be retained in memory.

After type provider disposal, any TypeProviderConfig objects provided to the
type provider will now raise ObjectDisposedException when the
SystemRuntimeContainsType method is called.

closes https://github.com/Microsoft/visualfsharp/pull/591
fixes https://github.com/Microsoft/visualfsharp/issues/106

commit d3b2f0e1b7f117de84feadcddc085a3cba307c09
Author: Don Syme <donsyme@fastmail.fm>
Date:   Fri Aug 14 14:35:53 2015 +0100

    testing for 2nd half of 106

commit adebe18e0291935d1b7eebd362abfef4b2dacc49
Author: Don Syme <donsyme@fastmail.fm>
Date:   Fri Aug 14 14:09:51 2015 +0100

    Second half of fix to https://github.com/Microsoft/visualfsharp/issues/106