Allow development on a freshly installed Windows 10
Remove hard-coded paths to SDK folder, and in one case, just update the list of SDKs to look for.

Fixes #561

The errors in #561 stemmed from inability to find the SDK utils, so strong-name validation was not turned off for the assemblies that are usually strong-named.

Fixed errors in config.bat