Send MSBuild output to AppVeyor
I did this because I was getting tired of waiting for entire
twenty-minute builds every time I wanted to see if I'd gotten the
test-result uploading right yet.

It didn't give me a whole lot to do while I waited, since it actually
worked the first time, but on the other hand that was also the time I
finally got the result uploading to work, so it didn't really matter.