An alternative List.filter implementation
This alternative implementation tries to reuse the tail of the list. It
does this first by filtering without copying whilst the predicate
returns true. If a false is found, the section that has been scanned is
then copied and it falls back to a more traditional building or result
list, with the exception that it keeps track of the final set of true
values in order to paste these to the end of the resultant list, thereby
reusing some of the original list, and freeing up the temporary tail for
garbage collection.