Simplified use of OnComplete & OnDispose
Pushed the chaining functionality into the interface and added extra
methods on SeqConsumer. This means the consumer can ignore the interface
and just implement their version, which means less likely to be an issue
of the message not being chained correctly.

It also has the advantage that in the object expressions we don't have
to cast back to the base type, which was a potentital area for errors.