Remove Helper from signature file
Probably shouldn't be exposed in that manor in the first place, but
secondly they caused a error in ci_part1. Used this as a chance to
rename the module as well.

Update LinqAggreagates output

Due to the change in the inline functions within Seq (i.e. sum etc.)
there are significant changes within this file.

Modified item/tryItem to use skip

Unit tests check that the item call object the lazy nature of the
Seq.init

Patch up the surface area. Still subject to change.

Starting the exposure of the inlinable Composer

- Still hidden via internal module
- simplified PipeIdx, no need for optional now
- Made ISeqFactory an abstract class instead of interface so as not to
require a stub implementation of PipeIdx in every object expression (or
alternatively if the abstract class was used with the interface, then
explicit declaration of the interface as well)
- filter and map changed to inline versions

Fix incorrect pipeIdx

Hack to stop tail calls on ICompletionChaining

passing a reference as an argument in a funciton stops the F# compiler
from outputting a tail instruction for that function.  None of these
functions will be significantly deep as to warrant the need for a tail
call.

mapi to inline version

- added a mapi_adapt version for non-inlined

Removed old choose function

localizing upto

This is retained for compatibility

cleaning up SeqComposer.Helpers

- better comments
- consistent casing

Seq.map3

Seq.mapi2

Simplified map2

- removing the check of both types

Seq.unfold

Added an IdentityFactory

Identity can be used to wrap basic containers into SeqComposer
compatible types, but can safely be removed when composing the
components.