Upgrade FsCheck to 2.6.2. (#1715)
Due to a change in FsCheck, tests run with a greater effective size
(i.e. generated lists, arrays, strings etc are significantly longer). To
make the tests complete in reasonable time the following changes were
made. First, the maximum size of the test runs in the
CollectionModulesConsistency and AsyncModule tests was changed to 25 and
20 resp. (down from the default, 50). Second, the `haveSameElements`
function was optimized so it is no longer quadratic. (Also note this
function contained a typo bug: `&& ys |>`, not xs.)