Don't append Module for types with arguments. (#1772)
Don't implicitly append 'Module' to module names if the type with the
same name as the module has a type argument. In that case the name is
not ambiguous.
In addition, the current situation breaks these fairly common use cases
for libraries that need to interoperate with other .NET languages, as
with the new compiler some modules will suddenly gain a Module suffix.
There is no way to turn this off, short of making the module into a
static class, which is silly.