Improve compiler's type searching logic (#1792)
The compiler has some logic to search for known types amongst available DLLs.

With this PR we get to enable several tests that uses F# Interactive on .NET Core and more tests for the F# compiler on .NET Core

Some of the known-type-search logic is OK, but swathes of it are not so good, e.g. this stuff

This improves the search by

just having one function to do the search

moving many "known types" out of Abstract IL (ILGlobals) and into the F# Compiler TcGlobals. THe advantages of this is that getting started with ILGlobals is much simpler if the set of referenced types is simpler - in particular if the small set of types known to Abstract IL are all always in the primary assembly (Int32, String, Array etc.).

This is particularly important for .NET Core. I don't trust the current code to correctly find some of the more obscure types for .NET Core, since the searching was based around various assumptions for Profile 7, 78, 259.

Overall this removes a fair chunk of fragile code from the compiler.

Separately this refactors TcGlobals to be a class instead of a record (which requires fewer lines of code)