Get correct options for scripts (#1865)
Prior to this, we never update the compiler options stored for scripts. This means that as you add #r or #load the compilation options aren't updated.

This PR updates the options in the semantic analysis phase. We can't update them on every request for options because it is too expensive.

This fixes some aspects of #1808 but the underlying problem of not reacting to file update events in dependencies is not fixed.