Surface area test output improvements.
Re-implemented the way the test which compares actual vs. expected
surface area works, so that when they don't match exactly, the output
only displays the members which don't match. The new output message
also shows which members were "unexpectedly present" and
"unexpectedly missing", which makes it easier to diagnose the cause
of the test failure.