Collapse Val with ValData, Entity with EntityData, Typar with TyparData (#2384)
The F# compiler code has always used a separation

ValRef --> Val --> ValData
EntityRef --> Entity --> EntityData
Typar --> TyparData
where the Val node was just a { mutable Data: ValData }. This was to enable the easy coding of the TastUnpickle deserialization routine which filled in the data (Link) while declaring ValData as immutable as possible. (likewise Entity, Typar)

The Val and ValData fields can be collapsed into one object if we accept that the Link routine can mutate the fields of a unchecked-initialized Val.

As well as saving the indirection during processing, the memory listings here indicate that about 5MB out of 150MB of data is spent on these Val/Tycon/Entity nodes.