Port build.cmd to build.sh
Ported majority of the functionality from build.cmd to build.sh
to make it easier to build/run the various test suites.
The usual 'net40' projects build just fine, but some of the projects in the
test suites seem to have build errors unrelated to build.sh.

Accept but ignore 'ci' target for now.

The mono distribution for Linux doesn't seem to include nuget as it does
on macOS. Try to use built-in nuget but fall back to the binary in the
repo if necessary.

Use the POSIX built-in 'command' instead of 'which'; the latter may not
be available on all platforms, and not all implementations have the same behavior.
(See dotnet/coreclr#8823 for a similar change.)