Follow-up to #2497. (#2536)
Fix FSharpSource.targets so ILComparer builds on mono.
In build.sh, preface invocations of nunit-console.exe with `mono`,
so it runs as expected. Move the net40-fsharpqa tests last,
unit tests (when they work) are faster than the tests in fsharpqa
so we'll fail faster when there's an obvious breaking change.