Implement script for diffing fsharp repos. (#2597)
Implement a shell script wrapping ``git diff`` to diff the fsharp/fsharp and
Microsoft/visualfsharp repositories and produce a patch. The patch is useful
when looking for changes which need to be synced (manually) between the two
repositories, and for changes in projects or build scripts that may result in
behaviorial differences between code compiled from the different repos.