First working interpolation for strings
Most of the added code is from the old PR of @forki. I copied it partly because I tried to better understand how it should work roughly.

However I think it's better to move a part of the parsing to the parser and lexer. But that's for later.

Do not try to interpolate integers, this WILL result in an Fatal-Execution-Engine-Error. I'm afraid.

Signed-off-by: realvictorprm <mueller.vpr@gmail.com>