Adjustments to TESTGUIDE.md relative to usage of fsharpqa tests
Tentative to add a .fsx test runner calling into perl, idea will be, once it is having tests pass, to make the script modify the test list file to tag arbitrary tests and run only those, and then revert the tagging. I'm not sure if this can be made to work consistently but it is worth trying.