Stop treating en as a localization target
For reasons that have been lost to the sands of time, F# treats the "en" locale as a localization target even though all the translatable resources start out in English. The build produces en resource assemblies, which are never used, while still embedding the original English resources in the primary assemblies for use as the neutral locale.

This is unnecessary and causes problems elsewhere. This commit removes en from the list of localization targets and deletes all the *.en.xlf files.