Consume localized resources
Consume the localized resources produced for menu items, buttons, etc. in FSharp.ProjectSystem.FSharp.dll. Because of conflicts caused by having en as both a localization target and being used as the "neutral" locale, localization of MenusAndCommands.vsct was turned off. We still had the associated .xlf files and the loc team was periodically updating them, but we never consumed their contents. As such, all the text from that file is unlocalized.

Now that we've stopped treating en as a localization target we can turn it back on. This does require adding an MSBuild target to work around a different issue (namely, the VSSDK really, really wants you to put *all* resources in separate assemblies).