Removing CONTRIBUTING.md
I think this document is outdated and duplicating too much information at this point. I also think most people don't read it or know it even exists. In an effort to simplify the amount of documents we either explicitly or implicitly expect people to read, I think this can be removed.