Async module is inconsistent in using SynchronizationContext
Some of the functions queue new computations on the thread pool while some on the synchronization context. When using a custom synchronization context (like epoll) this breaks the synchronization context model.

Task for example is consistent in using synchronization context.

This commit fixes it by first trying to post the continuation to the current synchronization context and only then fallback to the thread pool