Change type signature of IFileSystem.DirectoryCreateShim.
The only part of IFileSystem which remains coupled to the genuine
filesystem is DirectoryCreateShim, which currently returns a
DirectoryInfo. This commit changes it to return a string instead: the
FullName of that DirectoryInfo. After this commit, it is possible (at
least according to the types) to mock out the filesystem entirely.