Add PathCombineShim, and move some FileSystem functions to extension methods
Another place where the real filesystem leaks into FCS is in its liberal
use of Path.Combine. This commit shims out Path.Combine and uses the
shim in a number of places.

This commit also makes a couple of the IFileSystem functions into
extension methods, since they can be implemented in terms of functions
which are already present. This will make it easier for implementors to
create their own IFileSystem.