Recognize IsByRefLikeAttribute from user code. (#11374)
* Allow IsByRefLikeAttribute to be specified in user code or a non-framework assembly.

And use isByrefLikeTyconRef instead of specifically checking whether the attribute is applied.
A post-inference check about this attribute was moved to the aforementioned function.

* Fix a typo.

* Remove some leftover code that was causing errors.

And use tname_IsByRefLikeAttribute without the TcGlobals qualification.

* Handle the ILTypeSpec.Enclosing property correctly.

It wasn't holding the namespace's components but the inner types if it was nested. The project now builds.

* Test that custom IsByRefLikeAttributes are recognized.

* Fix one failing test.

* Fix another failing test.

* Address PR feedback: move a check back to PostInferenceChecks.fs.

* Add another test.