Remove need for passed in "ArcadeSdk" variable in source build (#15834)
Instead of using a passed in variable to determine where arcade lives, use a similar strategy of a few other repos (deployment tools, runtime, diagnostics) of using the parent directory of the tasks assembly, which is a property always available. This can then be used to find the release packages targets.
Avoids the need to pass in the arcade SDK location as an env variable for FSharp.