Initial look at issue meant to give priority to extension method if a property of same name exists and arguments are passed to it (assuming property is not a function value).
Lots of things to sort out, this just takes care of the first layer of interaction between CheckExpression and NameResolution.

Given impact on service / tooling, it is not clear that adding an entry in NameResolution.Item is the best choice, may reconsider into changing the return type of ResolveLongIdentInTypePrim instead.