feat: Add function to get return type of symbol
This commit adds a new function `GetReturnTypeOfSymbol` which takes a symbol name and a document as input, and returns the return type of the symbol. This function is used in the test case `Correctly infer int as explicit return type` to verify that the inferred return type of the symbol "sum" is "int".