revisit DeclarationListInfo.Create to avoid regression where we'd show both event and extension method (only event can be called).
The new logic 100% preserves the previous behaviour in absence of the particular condition we are interested for in RFC-1137.

In case the condition for RFC-1137 hits, we partition the items for those to be split apart (property and extension methods), taking care of discarding items as we yield the expected tuples that correspond to each item and the rest is processed as it was originally.

There are still probably loopholes with items showing up separated, but this would only impact if there are actually extension method and property with same name.