Remove unnecessary DotNetBuildOffline switches
DotNetBuildOffline is set in two test projects in the fsharp repo. Test projects are already filtered out when building from source, so those switches are unnecessary.

We plan to overhaul the source build control switches and remove this one completely. FSharp was the only repo that was still using it.