Nullness feature :: various bugfixes (#17080)
* Bugfix - matching aliased nullable should strip nullness

Eliminating nullness after pattern matching null (that is , for subsequent patterns) must visit contents of abbreviations as well. Otherwise it does not work with the Maybe<T> type whcih we use in the compiler.

* Making 'obj' work with new 'not null' constraints in fslib functions

Bugfix: obj cannot be passed to generic typars which require T: not null, such as the NonNull active pattern.

This commit fixes it.

* Bugfix - false 'useless null' warning in nested applications

Error fixed:
Error on useless null checkwith nullness constraint propagation in code like this:let meTry = Option.ofObj (Path.GetDirectoryName "")`. The warning about 'useless Option.ofObj' points to the string literal, ignoring the string literal is first passed to an API which may return null

* Fix import for C# extension methods

Bugfix for:
C# extension methods which put "?" on the this argument are wrongly interpreted by moving the nullability elsewhere. See AsMemory<T> from System.Memory.dll , this treats byteArray.ToMemory() as resulting in a Memory<byte | null> which is clearly wrong.

Also, this now allows to call C# extension methods with ?this to be invoked on a nullable value.

* LinkedList First,Last bugfix

There was a bug of LinkedList .First and .Last properties not returning nullable nodes.
This was fixed by improved byte import in previous commit, adding a regression test for guarding this.


* Bugfix: Solve nullness for typars

This fixes a bug where `not null` generic constraint was incorrectly passed between two typars:`T1 | null` with not null constraint on T1, and T2 without constraints.

This occured when calling Option.ofObj(..) when the inner expression caused solving of generic type arguments, e.g. after (|>) or (id) function.

This uses additional inference variable to unify them.