Do not throw from IsUnionCaseTester and handle a case where symbol is not considered a method/property (#17634)
* WIP some unit tests

* Return false instead of throwing

* Fix IsUnionCaseTester for generated code

Seems like methods/properties that are generated in IL, like `get_Is*`
in this case, have `IsMethod` (or `IsProperty`) false for some reason,
even when `IsPropertyGetterMethod` is true. This would result in
`IsUnionCaseTester` giving incorrect answers. This fixes that at
`IsUnionCaseTester`, though it might be worth it to see if it can be
fixed at the root of the issue

* Add a release note

* Move helpers around in unit tests for more reuse

* Add negative unit test

---------

Co-authored-by: Adam Boniecki <abonie@users.noreply.github.com>
Co-authored-by: Vlad Zarytovskii <vzaritovsky@hotmail.com>