Fix overbuilding fsharp & define solution to build in script
FSharp was building two solutions in the VMR on Windows which resulted
in the inner-build building the same thing but in parallel. The builds
each take 7-8 minutes. Fix that by introducing a productBuild switch
(which is already exposed in the Arcade scripts) and condition
which solution to build on it or the existing source-build switch.

Remove the /p:Projects property from DotNetBuild.props.