Treat `(|P|) : 'a -> 'b` similarly to `(|P|) : 'a`
That is, for an active pattern whose type is known to be a function type
but whose range type is not yet solved, assume that it _could_ be solved
to, e.g., another function type. This can show up in, e.g., an active
pattern that is defined mutually-recursively with `and`, has no type
annotations, and returns a lambda with `fun` or `function`.