Add new sln file for LSP proj on ubuntu
Create a solution file to be run by copilot on a non-windows machine.
Exclude Editor project from it due to windows dependencies. The LSP
project itself targets net8.0-windows - TODO investigate if this is
needed, we have to either conditionally change the target framework for
copilot or just drop "-windows" part