Run IL spread tests only against .NET (Core)
The compiler emits various types from the
System.Diagnostics.CodeAnalysis namespace for the .NET Framework target,
but those types come from the runtime for the .NET (Core) target.
Since the only IL that is material here is the field names, types, and
ordering, and since the spread logic is entirely
framework/runtime-agnostic, it is simpler to run these tests only for
the .NET (Core) target.